Feature/movella (#37)
* Added slip calculations and coordinate transforms for pose and pose dot.

* Bug fixes and local velocity.

* Bug fixes.

* DCM Slip and velocity logging.

* Velocity transform fixes.

* Fixes.

* Power limit logging.

* Fixed accel: linear feedforward schedule with slope 11.29m/s^2.

* Ignore gitcommit.h.

* Removed testing code.

* fix headers

---------

+++ stm32f413-drivers/CMR/include/CMR/fdcan.h
@@ -197,7 +197,7 @@ void cmr_canFilter(
 
 int cmr_canTX(
     cmr_can_t *can,
-    uint16_t id, const void *data, size_t len,
+    uint16_t id, const void *data, uint8_t len,
     TickType_t timeout
 );
